requirements.txt
torch
torchvision
torchaudio
timm
fastapi
uvicorn
python-multipart
albumentations
pillow
numpy